+++ src/hal/utils/halcmd_completion.c
@@ -136,7 +136,7 @@ static int writer_match(hal_pin_dir_t dir, int writers) {
 
 static void check_match_type_pin(const char *name) {
     int next = hal_data->pin_list_ptr;
-    int sz = strcspn(name, " \t");
+    size_t sz = strcspn(name, " \t");
 
     while(next) {
         hal_pin_t *pin = SHMPTR(next);
@@ -151,7 +151,7 @@ static void check_match_type_pin(const char *name) {
 
 static void check_match_type_signal(const char *name) {
     int next = hal_data->sig_list_ptr;
-    int sz = strcspn(name, " \t");
+    size_t sz = strcspn(name, " \t");
 
     while(next) {
         hal_sig_t *sig = SHMPTR(next);
